# Conflicts:
#	src/main/java/net/Indyuce/mmocore/api/player/profess/PlayerClass.java
#	src/main/java/net/Indyuce/mmocore/api/quest/trigger/ExperienceTrigger.java
#	src/main/java/net/Indyuce/mmocore/experience/ExperienceObject.java
#	src/main/java/net/Indyuce/mmocore/experience/Profession.java
#	src/main/java/net/Indyuce/mmocore/experience/SimpleExperienceObject.java

+++ src/main/java/net/Indyuce/mmocore/listener/option/RedirectVanillaExp.java
@@ -16,7 +16,7 @@ public class RedirectVanillaExp implements Listener {
 
 	@EventHandler
 	public void a(PlayerExpChangeEvent event) {
-		int a = (int) (event.getAmount() * ratio);
+		double a = (event.getAmount() * ratio);
 		if (a > 0)
 			PlayerData.get(event.getPlayer()).giveExperience(a, EXPSource.VANILLA);
 	}
